% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghettoblaster.R
\name{blast}
\alias{blast}
\title{Run a BLAST search}
\usage{
blast(sequence, max_num_seq = 100, verbose = FALSE, add_worms = FALSE)
}
\arguments{
\item{sequence}{A DNA sequence}

\item{max_num_seq}{The maximum number of sequences to return}

\item{verbose}{Whether to print status messages}

\item{add_worms}{Whether to add WoRMS taxonomy information}
}
\value{
A data frame with the BLAST results
}
\description{
Run a BLAST search
}
