% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghettoblaster.R
\name{get_results}
\alias{get_results}
\title{Get the results of a BLAST request}
\usage{
get_results(rid)
}
\arguments{
\item{rid}{A request ID}
}
\value{
A data frame with the results
}
\description{
Get the results of a BLAST request
}
