% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghettoblaster.R
\name{submit_sequence}
\alias{submit_sequence}
\title{Submit a sequence to NCBI BLAST}
\usage{
submit_sequence(sequence, program = "blastn", max_num_seq = 100)
}
\arguments{
\item{sequence}{A DNA sequence}

\item{program}{The BLAST program to use}

\item{max_num_seq}{The maximum number of sequences to return}
}
\value{
A request ID
}
\description{
Submit a sequence to NCBI BLAST
}
