% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghettoblaster.R
\name{wait_until_ready}
\alias{wait_until_ready}
\title{Wait until a BLAST request is ready}
\usage{
wait_until_ready(rid, verbose = FALSE)
}
\arguments{
\item{rid}{A request ID}

\item{verbose}{Whether to print status messages}
}
\value{
TRUE if the request is ready, FALSE in case of an unexpected status
}
\description{
Wait until a BLAST request is ready
}
