% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghettoblaster.R
\name{worms_for_names}
\alias{worms_for_names}
\title{Get WoRMS taxonomy for a list of species names}
\usage{
worms_for_names(...)
}
\arguments{
\item{species_names}{A character vector of species names}
}
\value{
A data frame with the taxonomy, input names in column input
}
\description{
Get WoRMS taxonomy for a list of species names
}
